+++ tests/test_examples.py
@@ -20,6 +20,8 @@ for dirpath, dirnames, filenames in os.walk(example_folder):
 
 @pytest.mark.parametrize("path", examples)
 def test_example(path):
+    if "pickling" in path:
+        pytest.skip("Example requires to be executed as a top level module")
     with open(path, "r") as f:
         # If exec gets two separate objects as globals and locals, the code
         # will be executed as if it were embedded in a class definition, which